fix: drill tab change handler to sidestep react-router upgrade
